# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ipoh Drum Laravel API",
        "description": "This is a swagger-generated API documentation for the project Ipoh Drum. (Only supports OpenAPI Annotations for now.)",
        "contact": {
            "email": "henry_lcz97@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "The URL that IpohDrum Laravel API is running on."
        }
    ],
    "paths": {
...
        "/api/type": {
            "get": {
                "tags": [
                    "TypeControllerService"
                ],
                "summary": "Get list of types",
                "description": "Returns list of types",
                "operationId": "getTypeList",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of types"
                    },
                    "default": {
                        "description": "Unable to retrieve list of types"
                    }
                }
            },
            "post": {
                "tags": [
                    "TypeControllerService"
                ],
                "summary": "Creates a type.",
                "operationId": "createType",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Type name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Type Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "description": "Icon",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type has been created successfully."
                    },
                    "default": {
                        "description": "Unable to create the type."
                    }
                }
            }
        },
...
        "/api/filter/type": {
            "get": {
                "tags": [
                    "TypeControllerService"
                ],
                "summary": "Filter list of types",
                "description": "Returns list of filtered types",
                "operationId": "filterTypeList",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromdate",
                        "in": "query",
                        "description": "From Date for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "todate",
                        "in": "query",
                        "description": "To string for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status for filter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of filtered types"
                    },
                    "default": {
                        "description": "Unable to retrieve list of types"
                    }
                }
            }
        },
...
        "/api/type/{uid}": {
            "get": {
                "tags": [
                    "TypeControllerService"
                ],
                "summary": "Retrieves type by Uid.",
                "operationId": "getTypeByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Type_ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type has been retrieved successfully."
                    },
                    "default": {
                        "description": "Unable to retrieve the type."
                    }
                }
            },
            "put": {
                "tags": [
                    "TypeControllerService"
                ],
                "summary": "Update type by Uid.",
                "operationId": "updateTypeByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Type_ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Type name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Type Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "description": "Icon",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type has been updated successfully."
                    },
                    "default": {
                        "description": "Unable to update the type."
                    }
                }
            },
            "delete": {
                "tags": [
                    "TypeControllerService"
                ],
                "summary": "Set type's 'status' to 0.",
                "operationId": "deleteTypeByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Type ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type has been 'deleted' successfully."
                    },
                    "default": {
                        "description": "Unable to 'delete' the type."
                    }
                }
            }
        },
...
    },
...
}